Factor out info and review component so that we can adjust order for mobile
