+++ src/components/BottomBar/BottomBarCourse.vue
@@ -1,146 +1,17 @@
 <template>
   <div class="bottombarcourse">
     <div class="bottombarcourse-wrapper">
-      <div
-        :class="
-          isSmallerWidth
-            ? 'bottombarcourse-bar-info-noOverflow'
-            : 'bottombarcourse-bar-info-overflow'
-        "
-      >
-        <div class="info">
-          <div class="info-section-wrapper">
-            <div>
-              <div class="section">
-                <h1 class="info-head">Credits</h1>
-                <p class="info-fact">{{ courseObj.credits }}</p>
-              </div>
-              <div class="section">
-                <h1 class="info-head">Offered</h1>
-                <p class="info-fact">{{ courseSemesters }}</p>
-              </div>
-            </div>
-            <div>
-              <div class="section">
-                <h1 class="info-head">Instructors</h1>
-                <p class="info-fact">{{ courseInstructors }}</p>
-              </div>
-              <div class="section">
-                <h1 class="info-head">Enrollment Information</h1>
-                <p class="info-fact">{{ courseEnrollment }}</p>
-              </div>
-            </div>
-            <div>
-              <div class="section">
-                <h1 class="info-head">Distribution Category</h1>
-                <p
-                  class="info-fact"
-                  v-for="distributionCategory in courseDistributions"
-                  :key="distributionCategory"
-                >
-                  {{ distributionCategory }}
-                </p>
-              </div>
-              <div class="section">
-                <h1 class="info-head">{{ courseObj.lastRoster }} Lecture Information</h1>
-                <p
-                  class="info-fact"
-                  v-for="latestLecInfo in courseLectureTimes"
-                  :key="latestLecInfo"
-                >
-                  {{ latestLecInfo }}
-                </p>
-              </div>
-            </div>
-          </div>
-          <div class="info-link">
-            <a :href="rosterLink" class="info-link-blue" target="_blank">
-              View Course Information on Roster
-              <span class="info-link-blue-img"
-                ><img src="@/assets/images/link-blue.svg" alt="link arrow"
-              /></span>
-            </a>
-          </div>
-        </div>
+      <div class="bottombarcourse-bar-info-overflow" v-if="!isSmallerWidth">
+        <bottom-bar-course-info :courseObj="courseObj" />
       </div>
-      <div
-        :class="
-          isSmallerWidth
-            ? 'bottombarcourse-bar-details-noOverflow'
-            : 'bottombarcourse-bar-details-overflow'
-        "
-      >
-        <div class="details">
-          <div class="details-ratings-link-wrapper">
-            <a :href="CURLink" class="details-ratings-link" target="_blank">See All Reviews</a>
-          </div>
-          <div class="details-ratings-wrapper">
-            <div class="details-ratings">
-              <p class="details-ratings-title">
-                <span class="details-ratings-title-strong">Overall: </span>
-                <span class="details-ratings-strong">{{ CUROverallRating }}</span>
-              </p>
-              <div class="progress rating">
-                <div
-                  class="progress-bar"
-                  role="progressbar"
-                  :style="{
-                    width: `${(courseObj.overallRating / 5) * 100}%`,
-                    background: reviewsColor(courseObj.overallRating),
-                  }"
-                  aria-label="Overall Rating"
-                  :aria-valuenow="(courseObj.overallRating / 5) * 100"
-                  aria-valuemin="0"
-                  aria-valuemax="100"
-                ></div>
-              </div>
-            </div>
-            <div class="details-ratings">
-              <p class="details-ratings-title">
-                <span class="details-ratings-title-strong">Difficulty: </span>
-                <span class="details-ratings-strong"> {{ CURDifficulty }}</span>
-              </p>
-              <div class="progress rating">
-                <div
-                  class="progress-bar"
-                  role="progressbar"
-                  :style="{
-                    width: `${(courseObj.difficulty / 5) * 100}%`,
-                    background: reviewsColor(courseObj.difficulty, true),
-                  }"
-                  aria-label="Difficulty Rating"
-                  :aria-valuenow="(courseObj.difficulty / 5) * 100"
-                  aria-valuemin="0"
-                  aria-valuemax="100"
-                ></div>
-              </div>
-            </div>
-            <div class="details-ratings">
-              <p class="details-ratings-title">
-                <span class="details-ratings-title-strong">Workload: </span>
-                <span class="details-ratings-strong">{{ CURWorkload }}</span>
-              </p>
-              <div class="progress rating">
-                <div
-                  class="progress-bar"
-                  role="progressbar"
-                  :style="{
-                    width: `${(courseObj.workload / 5) * 100}%`,
-                    background: reviewsColor(courseObj.workload, true),
-                  }"
-                  aria-label="Workload Rating"
-                  :aria-valuenow="(courseObj.workload / 5) * 100"
-                  aria-valuemin="0"
-                  aria-valuemax="100"
-                ></div>
-              </div>
-            </div>
-          </div>
-          <div class="details-head">Prerequisites</div>
-          <p class="info-fact">{{ coursePrereqs }}</p>
-          <div class="details-head">Description</div>
-          <p class="info-fact">{{ courseObj.description }}</p>
-        </div>
+      <div class="bottombarcourse-bar-details-overflow" v-if="!isSmallerWidth">
+        <bottom-bar-course-review :courseObj="courseObj" />
+      </div>
+      <div class="bottombarcourse-bar-details-noOverflow" v-if="isSmallerWidth">
+        <bottom-bar-course-review :courseObj="courseObj" />
+      </div>
+      <div class="bottombarcourse-bar-info-noOverflow" v-if="isSmallerWidth">
+        <bottom-bar-course-info :courseObj="courseObj" />
       </div>
     </div>
   </div>
@@ -148,35 +19,11 @@
 
 <script lang="ts">
 import Vue, { PropType } from 'vue';
-import { reviewColors } from '@/assets/constants/colors';
-
-const joinOrNAString = (arr: readonly string[]): string =>
-  arr.length !== 0 && arr[0] !== '' ? arr.join(', ') : 'N/A';
-
-const noneIfEmpty = (str: string): string => (str && str.length !== 0 ? str : 'None');
-
-const naIfEmptyStringArray = (arr: readonly string[]): readonly string[] =>
-  arr && arr.length !== 0 && arr[0] !== '' ? arr : ['N/A'];
-
-const cleanCourseDistributionsArray = (distributions: readonly string[]): readonly string[] => {
-  // Iterates over distributions array and cleans every entry
-  // Removes stray parentheses, spaces, and commas
-  let matches: string[] = [];
-  if (distributions[0] === '') {
-    matches = ['N/A'];
-  } else {
-    for (let i = 0; i < distributions.length; i += 1) {
-      distributions[i].replace(/[A-Za-z0-9-]+/g, d => {
-        matches.push(d);
-        return d;
-      });
-    }
-  }
-
-  return matches;
-};
+import BottomBarCourseInfo from '@/components/BottomBar/BottomBarCourseInfo.vue';
+import BottomBarCourseReview from '@/components/BottomBar/BottomBarCourseReview.vue';
 
 export default Vue.extend({
+  components: { BottomBarCourseInfo, BottomBarCourseReview },
   data() {
     return {
       isSmallerWidth: window.innerWidth <= 976,
@@ -195,70 +42,6 @@ export default Vue.extend({
   methods: {
     isSmallerWidthEventHandler() {
       this.isSmallerWidth = window.innerWidth <= 976;
-    },
-
-    joinIfExists(arr: readonly string[]) {
-      return arr ? arr.join(',') : '';
-    },
-
-    reviewsColor(review: number, flip = false) {
-      const colors = Object.values(reviewColors);
-      let index;
-      if (review < 2) {
-        index = 0;
-      } else if (review >= 2 && review < 4) {
-        index = 1;
-      } else {
-        index = 2;
-      }
-
-      return flip ? colors[colors.length - 1 - index] : colors[index];
-    },
-  },
-
-  computed: {
-    courseSemesters(): string {
-      return joinOrNAString(this.courseObj.semesters);
-    },
-    courseInstructors(): string {
-      return joinOrNAString(this.courseObj.instructors);
-    },
-    courseEnrollment(): string {
-      return joinOrNAString(this.courseObj.enrollment);
-    },
-    coursePrereqs(): string {
-      return noneIfEmpty(this.courseObj.prereqs);
-    },
-    courseLectureTimes(): readonly string[] {
-      return naIfEmptyStringArray(this.courseObj.lectureTimes);
-    },
-    courseSubjectAndNumber(): { readonly subject: string; readonly number: string } {
-      const [subject, number] = this.courseObj.code.split(' ');
-      return { subject, number };
-    },
-    courseDistributions(): readonly string[] {
-      return cleanCourseDistributionsArray(this.courseObj.distributions);
-    },
-    rosterLink(): string {
-      return `https://classes.cornell.edu/browse/roster/${this.courseObj.lastRoster}/class/${this.courseSubjectAndNumber.subject}/${this.courseSubjectAndNumber.number}`;
-    },
-    CURLink(): string {
-      return `https://www.cureviews.org/course/${this.courseSubjectAndNumber.subject}/${this.courseSubjectAndNumber.number}`;
-    },
-    CUROverallRating() {
-      if (this.courseObj.overallRating === 0) return '';
-      if (!this.courseObj.overallRating) return 'N/A';
-      return Math.round(this.courseObj.overallRating * 10) / 10;
-    },
-    CURDifficulty() {
-      if (this.courseObj.difficulty === 0) return '';
-      if (!this.courseObj.difficulty) return 'N/A';
-      return Math.round(this.courseObj.difficulty * 10) / 10;
-    },
-    CURWorkload() {
-      if (this.courseObj.workload === 0) return '';
-      if (!this.courseObj.workload) return 'N/A';
-      return Math.round(this.courseObj.workload * 10) / 10;
     },
   },
 });
@@ -304,114 +87,6 @@ export default Vue.extend({
       }
     }
   }
-}
-
-.info {
-  margin: 20px;
-  display: flex;
-  flex-direction: column;
-  height: 90%;
-  justify-content: space-between;
-  margin-bottom: 0rem;
-
-  &-head {
-    margin-top: 5px;
-    font-weight: 500;
-    font-size: 12px;
-    color: $darkGray2;
-    margin-bottom: 0rem;
-  }
-
-  &-fact {
-    padding-right: 10px;
-    font-size: 16px;
-    color: $primaryGray;
-    margin-bottom: 0rem;
-  }
-
-  &-link {
-    font-size: 16px;
-    line-height: 16px;
-    text-decoration-line: underline;
-    margin-top: inherit;
-
-    &-blue {
-      color: $yuxuanBlue;
-      // TODO: update picture
-      font-weight: 500;
-
-      &-img {
-        margin-left: 0.2rem;
-      }
-    }
-  }
-}
-
-.section {
-  margin-bottom: 8px;
-  float: left;
-  width: 50%;
-  height: inherit;
-}
-
-.details {
-  margin: 20px;
-
-  &-head {
-    margin-top: 10px;
-    font-weight: 500;
-    font-size: 12px;
-    color: $darkGray2;
-    margin-bottom: 0rem;
-  }
-
-  &-ratings {
-    margin-right: 3%;
-    width: 30%;
-    float: left;
-    margin-bottom: 1rem;
-
-    &-title {
-      font-size: 16px;
-      line-height: 16px;
-      color: $primaryGray;
-
-      &-strong {
-        font-weight: 500;
-      }
-    }
-
-    &-strong {
-      font-weight: 2000;
-    }
-
-    &-link {
-      font-size: 14px;
-      text-decoration-line: underline;
-      color: $yuxuanBlue;
-      &-wrapper {
-        display: flex;
-        flex-direction: row-reverse;
-        width: 100%;
-        margin-top: -1rem;
-        margin-bottom: 1rem;
-        margin-left: 1rem;
-      }
-    }
-  }
-}
-
-.progress-bar {
-  transition: width 0s ease;
-}
-
-.rating {
-  width: 100%;
-  border-radius: 100px;
-}
-
-.hide {
-  height: 0;
 }
 
 @media only screen and (max-width: 976px) {
@@ -461,16 +136,6 @@ export default Vue.extend({
   .bottombarcourse {
     left: 0rem;
     width: 100%;
-  }
-  .details {
-    &-ratings {
-      width: 60%;
-      &-wrapper {
-        display: flex;
-        flex-direction: column;
-        width: 100%;
-      }
-    }
   }
 }
